requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
anyio==3.7.1
apischema==0.16.6
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.1.0
base58==2.1.1
based58==0.1.1
bitarray==3.1.0
cachetools==4.2.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
construct==2.10.68
construct-typing==0.5.6
crc32c==2.7.1
cryptography==44.0.1
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
OSlash==0.6.3
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.5
pytest-asyncio==0.25.3
python-telegram-bo.t==21.10
referencing==0.36.0
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.22.3
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
soldier==0.2.5
SQLAlchemy==2.0.38
tenacity==9.0.0
tonsdk==1.0.15
tvm-valuetypes==0.0.12
types-cachetools==4.2.10
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
websockets==10.4
yarl==1.18.3
